+++ src/app/AuthSession.js
