+++ src/app/api/detail/route.js
