+++ src/app/api/list/[...param]/route.js
