+++ src/app/detail/[id]/detail.module.css
