+++ src/app/detail/[id]/page.js
