+++ src/app/test/result/[type]/page.js
