+++ src/components/button/LikeBtn.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useSession } from "next-auth/react";
-import Modal from "@/app/components/common/Modal";
+import Modal from "../common/Modal";
 import { HiOutlineHeart } from "react-icons/hi2";
 import { HiMiniHeart } from "react-icons/hi2";
